Implement saving and loading the game
This saves the player's progress to /var/phoenix/saved-game, and loads
it when you start again. The UX for this feature isn't great in the
original Phoenix, though - it'd be nice to give the player a chance to
decide to resume an existing save or start a new game from the main
menu at some point.

Closes #203