Store unrelocated image addresses for game objects
There was a problem with saving and loading the game because the game
state would be saved with absolute image addresses. The next time you
load the game, Phoenix could be running from somewhere else in memory.
This change replaces the absolute addresses with relative ones and
relocates them on demand.